makediff: look in subjects only
We currently use rev-list with -F to find commits.
This helps if commit subject has any special characters
but means that pattern can apply anywhere in the line.
In particular, a common Fixes: (hash) "subject" pattern
we started using implies that fixups are picked up
instead of the commit.

For now, just switch to using regexp (dropping -F)
and add "^" in front to only look for the string at
beginning of the line.

We'll worry about more elaborate logic later, if it's ever
needed.

Message-Id: <1130700f4ae71a93d9887af35cf105731219f057.1721062136.git.mst@redhat.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>