virtio-gpu: support blob alignment information
There's an increasing number of machines supporting multiple page sizes
and, on these machines, the host and a guest can be running with
different pages sizes. In addition to this, there might be physical
devices that require to operate their memory at a specific granularity.

In these cases, if they are to use Shared Memory Regions, the device
and the driver must operate with the same granularity, as otherwise
the former might not be able to fulfill the requests sent by the
latter.

For the GPU device, this has an impact on blob creation and mapping. To
address the problem, allow the device to require certain alignment
constrains for blob resources by extending the device configuration
with the field "blob_alignment" and introducing the
VIRTIO_GPU_F_BLOB_ALIGNMENT feature.

Signed-off-by: Sergio Lopez <slp@redhat.com>