virtio-rtc: Add initial device specification
The virtio-rtc device provides information about current time through
one or more clocks. As such, it is a Real-Time Clock (RTC) device.

The normative statements for this device follow in the next patch.

For this device, there is a Linux kernel driver patch series which is
being upstreamed, and a proprietary device implementation.

Miscellaneous
-------------

The spec does not specify how a driver should interpret clock readings,
esp. also not how to perform clock synchronization.

The device uses the former "Timer/Clock" device id which is already part
of the specification. This device id was registered a long time ago and
should be unused according to the author's information. The name "RTC"
was determined to be the best for a device which focuses on current
time.

Signed-off-by: Peter Hilber <quic_philber@quicinc.com>
Link: https://lore.kernel.org/r/20250710090648.1711-3-quic_philber@quicinc.com